# backend/docs/swagger.yaml
definitions:
  UserCredentials:
    properties:
      password:
        example: daniel321
        minLength: 6
        type: string
      username:
        example: daniel
        type: string
    required:
    - password
    - username
    type: object
  UserSigninResponse:
    properties:
      id:
        type: integer
    type: object
info:
  contact: {}
  description: Backend API sepcifications for dancord
  title: dancord API
  version: "0.1"
paths:
  /ping:
    get:
      consumes:
      - application/json
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Ping the server
      tags:
      - ping
  /user/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password must be atleast 6 characters.
        in: body
        name: SignInBody
        required: true
        schema:
          $ref: '#/definitions/UserCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserSigninResponse'
        "400":
          description: Invalid Credentials
      summary: Sign a user into dancord
      tags:
      - user
  /user/signup:
    post:
      consumes:
      - application/json
      description: Insert description here
      parameters:
      - description: Password must be atleast 6 characters.
        in: body
        name: SignUpBody
        required: true
        schema:
          $ref: '#/definitions/UserCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserSigninResponse'
      summary: Sign a user up to dancord
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    description: Token in Bearer format to authenticate the user
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"